raamatud/requirements.txt
flask==3.0.2
requests==2.31.0
azure-storage-blob
flask-cors